+++ pages/index.tsx
@@ -1,5 +1,8 @@
 import { AddUrlBox } from "../src/components/AddUrlBox";
 import { Layout } from "../src/components/Layout";
+import { UrlCards } from "../src/components/UrlCards";
+import { url_card_data } from "../src/data/mockData";
+// add it to firebase & apis
 import type { NextPage } from "next";
 import Head from "next/head";
 import Image from "next/image";
@@ -8,13 +11,13 @@ const Home: NextPage = () => {
   return (
     <Layout>
       {/* SEARCH BOX */}
-      <div className="p-5 text-center text-white p2 bg-slate-900">
+      <div className="w-4/5 p-10 m-5 text-center rounded-md text-slate-900">
         Search box
       </div>
       {/* ADD BUTTON */}
       <AddUrlBox />
       {/* URL CARDS */}
-      <div>Url Cards 1</div>
+      <UrlCards cardsData={url_card_data} />
       <div>Url Cards 2</div>
     </Layout>
   );